refactor(template): remove local template execution tests from validation tests
